chore: fix silent error discards and clear 21 ESLint errors
Backend:
- checks.go: actually log the event create failure that the comment promised to log
- docker_discovery.go: log rollback App.Delete failures so orphaned rows are visible

Frontend (lint was failing with 21 errors, blocking the PR checklist):
- Extract CheckForm helpers to checkFormHelpers.tsx (react-refresh/only-export-components)
- Remove export from getSourceName in EventRow.tsx — only used internally
- Add eslint-disable for useAutoRefresh in AutoRefreshContext.tsx — conventional context pattern
- Wrap setState-calling useEffect bodies in async IIFEs across EventFeed, Dashboard, Events, ProxmoxDetail, Settings, TraefikDetail (react-hooks/set-state-in-effect)
- Fix Date.now() in ProxmoxDetail filter via useMemo (react-hooks/purity)
- Fix regex literal spaces in AppTemplateEditor.tsx (no-regex-spaces)
- Fix bare ternary expressions in Topology.tsx (no-unused-expressions)
