+++ frontend/src/components/checkFormHelpers.tsx
@@ -1,0 +1,159 @@
+import type { ReactNode } from 'react'
+import type { MonitorCheck, CreateCheckInput, CheckType, SSLSource } from '../api/types'
+
+// ── Types ─────────────────────────────────────────────────────────────────────
+
+export type FormFields = {
+  type: CheckType
+  name: string
+  target: string
+  interval_secs: string
+  expected_status: string
+  ssl_warn_days: string
+  ssl_crit_days: string
+  ssl_source: SSLSource
+  integration_id: string
+  traefik_domain: string
+  skip_tls_verify: string  // 'true' | 'false'
+}
+
+export const defaultForm: FormFields = {
+  type: 'ping',
+  name: '',
+  target: '',
+  interval_secs: '60',
+  expected_status: '200',
+  ssl_warn_days: '30',
+  ssl_crit_days: '7',
+  ssl_source: 'standalone',
+  integration_id: '',
+  traefik_domain: '',
+  skip_tls_verify: 'false',
+}
+
+// ── Helpers ───────────────────────────────────────────────────────────────────
+
+export function validateForm(form: FormFields): string | null {
+  if (!form.name.trim()) return 'Name is required'
+  const interval = parseInt(form.interval_secs, 10)
+  if (isNaN(interval) || interval < 30) return 'Interval must be at least 30 seconds'
+
+  if (form.type === 'url') {
+    if (!form.target.trim()) return 'Target is required'
+    if (!form.target.startsWith('http://') && !form.target.startsWith('https://')) {
+      return 'Target must begin with http:// or https://'
+    }
+  }
+
+  if (form.type === 'ssl') {
+    if (form.ssl_source === 'traefik') {
+      if (!form.traefik_domain) return 'Select a domain from the Traefik cert list'
+    } else {
+      if (!form.target.trim()) return 'Target is required'
+      if (!form.target.startsWith('http://') && !form.target.startsWith('https://')) {
+        return 'Target must begin with http:// or https://'
+      }
+    }
+  }
+
+  if (form.type === 'ping' && !form.target.trim()) return 'Target is required'
+
+  return null
+}
+
+export function checkToForm(check: MonitorCheck): FormFields {
+  return {
+    type: check.type as CheckType,
+    name: check.name,
+    target: check.type === 'ssl' && check.ssl_source === 'traefik' ? '' : check.target,
+    interval_secs: String(check.interval_secs),
+    expected_status: String(check.expected_status ?? 200),
+    ssl_warn_days: String(check.ssl_warn_days ?? 30),
+    ssl_crit_days: String(check.ssl_crit_days ?? 7),
+    ssl_source: (check.ssl_source as SSLSource) ?? 'standalone',
+    integration_id: check.integration_id ?? '',
+    traefik_domain: check.ssl_source === 'traefik' ? check.target : '',
+    skip_tls_verify: check.skip_tls_verify ? 'true' : 'false',
+  }
+}
+
+export function formToInput(form: FormFields, integrationID?: string): CreateCheckInput {
+  const input: CreateCheckInput = {
+    name: form.name.trim(),
+    type: form.type,
+    target: form.type === 'ssl' && form.ssl_source === 'traefik'
+      ? form.traefik_domain
+      : form.target.trim(),
+    interval_secs: parseInt(form.interval_secs, 10),
+  }
+  if (form.type === 'url') {
+    input.expected_status = parseInt(form.expected_status, 10)
+    input.skip_tls_verify = form.skip_tls_verify === 'true'
+  }
+  if (form.type === 'ssl') {
+    input.ssl_warn_days = parseInt(form.ssl_warn_days, 10)
+    input.ssl_crit_days = parseInt(form.ssl_crit_days, 10)
+    input.ssl_source = form.ssl_source
+    if (form.ssl_source === 'traefik' && integrationID) {
+      input.integration_id = integrationID
+    }
+  }
+  return input
+}
+
+export function renderCheckResult(check: MonitorCheck): ReactNode {
+  if (!check.last_result) return <span className="check-result-empty">No result recorded yet</span>
+  try {
+    const r = JSON.parse(check.last_result) as Record<string, unknown>
+    if (check.type === 'ping') {
+      return (
+        <div className="check-result-grid">
+          <span className="check-result-label">Latency</span>
+          <span className="check-result-value">{r.latency_ms != null ? `${r.latency_ms}ms` : '—'}</span>
+        </div>
+      )
+    }
+    if (check.type === 'url') {
+      return (
+        <div className="check-result-grid">
+          <span className="check-result-label">HTTP Status</span>
+          <span className="check-result-value">{r.status_code != null ? String(r.status_code) : '—'}</span>
+          <span className="check-result-label">Latency</span>
+          <span className="check-result-value">{r.latency_ms != null ? `${r.latency_ms}ms` : '—'}</span>
+          {!!r.error && <>
+            <span className="check-result-label">Error</span>
+            <span className="check-result-value check-result-error">{String(r.error)}</span>
+          </>}
+        </div>
+      )
+    }
+    if (check.type === 'ssl') {
+      if (r.error) return (
+        <div className="check-result-grid">
+          <span className="check-result-label">Error</span>
+          <span className="check-result-value check-result-error">{String(r.error)}</span>
+        </div>
+      )
+      const expiresStr = r.expires_at
+        ? new Date(r.expires_at as string).toLocaleDateString(undefined, { year: 'numeric', month: 'short', day: 'numeric' })
+        : '—'
+      return (
+        <div className="check-result-grid">
+          <span className="check-result-label">Days remaining</span>
+          <span className="check-result-value">{r.days_remaining != null ? String(r.days_remaining) : '—'}</span>
+          <span className="check-result-label">Expires</span>
+          <span className="check-result-value">{expiresStr}</span>
+          {!!r.issuer && <>
+            <span className="check-result-label">Issuer</span>
+            <span className="check-result-value">{String(r.issuer)}</span>
+          </>}
+          {!!r.subject && <>
+            <span className="check-result-label">Subject</span>
+            <span className="check-result-value">{String(r.subject)}</span>
+          </>}
+        </div>
+      )
+    }
+  } catch { /* fall through */ }
+  return <span className="check-result-empty">{check.last_result}</span>
+}
