+++ frontend/src/context/AutoRefreshContext.tsx
@@ -41,6 +41,7 @@ export function AutoRefreshProvider({ children }: { children: React.ReactNode })
   )
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
 export function useAutoRefresh(): AutoRefreshContextValue {
   return useContext(AutoRefreshContext)
 }
